refactor tier 1 subclass mapping method
