+++ src/retriever/server.py
@@ -16,6 +16,7 @@ from retriever.config.general import CONFIG
 from retriever.config.logger import configure_logging
 from retriever.config.openapi import OPENAPI_CONFIG, TRAPI
 from retriever.data_tiers import tier_manager
+from retriever.lookup.subclass import SubclassMapping
 from retriever.lookup.utils import QueryDumper
 from retriever.metadata.optable import OP_TABLE_MANAGER
 from retriever.query import get_job_state, make_query
@@ -54,12 +55,14 @@ async def lifespan(_: FastAPI) -> AsyncGenerator[None]:
     query_dumper = QueryDumper()
     if CONFIG.tier0.dump_queries or CONFIG.tier1.dump_queries:
         query_dumper.initialize()
+    await SubclassMapping().initialize()
 
     yield  # Separates startup/shutdown phase
 
     # Shutdown
     if query_dumper.initialized:
         query_dumper.wrapup()
+    await SubclassMapping().wrapup()
     await tier_manager.close_drivers()
     await OP_TABLE_MANAGER.wrapup()
     await REDIS_CLIENT.close()
